+++ __tests__/expect/toExist.spec.js
@@ -22,6 +22,18 @@ it("doesn't throw if expectation is met", async () => {
     .resolves.toEqual(undefined);
 });
 
+it("inverses the expectation when used with .not", async () => {
+  const ref = createFindElementMock();
+
+  jest.spyOn(appiumService, "findElement").mockResolvedValue(ref);
+  jest.spyOn(appiumService, "getElementExists").mockResolvedValue(false);
+
+  const $element = await element(by.label("button"));
+
+  await expect(assert($element).not.toExist())
+    .resolves.toEqual(undefined);
+});
+
 it("throws if expectation is not met", async () => {
   const ref = createFindElementMock();
 
@@ -36,6 +48,23 @@ it("throws if expectation is not met", async () => {
   } catch (err) {
     expect(err).toBeInstanceOf(Error);
     expect(err).toHaveProperty("message", "Expected element to exist.");
+  }
+});
+
+it("throws if expectation is not met when used with .not", async () => {
+  const ref = createFindElementMock();
+
+  jest.spyOn(appiumService, "findElement").mockResolvedValue(ref);
+  jest.spyOn(appiumService, "getElementExists").mockResolvedValue(true);
+  expect.assertions(2);
+
+  try {
+    const $element = await element(by.label("button"));
+
+    await assert($element).not.toExist();
+  } catch (err) {
+    expect(err).toBeInstanceOf(Error);
+    expect(err).toHaveProperty("message", "Expected element not to exist.");
   }
 });