+++ src/worker/services/appiumService.js
@@ -387,7 +387,6 @@ function createAppiumService(sessionStore) {
   // ({ sessionId: String?, element: AppiumElement, options? Object }) => Promise<String>.
   const getElementText = ({ sessionId = sessionStore.getSessionId(), element, options = {} }) => {
     if (!options.recursive) {
-      console.log("simple get text");
       return getElementTextAttribute({ sessionId, element })
     }
 
@@ -553,6 +552,14 @@ function createAppiumService(sessionStore) {
     });
   };
 
+  // ({ sessionId: String?, element: AppiumElement }) => Promise<String>.
+  const takeElementScreenshot = ({ sessionId = sessionStore.getSessionId(), element }) => {
+    return request({
+      method: "GET",
+      path: `/session/${sessionId}/element/${element.ELEMENT}/screenshot`
+    });
+  };
+
   return {
     getStatus,
     createSession,
@@ -592,7 +599,8 @@ function createAppiumService(sessionStore) {
     longPressElement,
     swipeElement,
     sendElementKeys,
-    clearElementText
+    clearElementText,
+    takeElementScreenshot
   };
 }